manager: Ensure strict ordering of setters (fix #2)
Many setters had issues where the messages could be received out of order, but
were processed in the order they were received, as opposed to the order they
were sent. This adds additional validation and discards setter calls that
aren't the most recent ones from a given sender. For now this is only applied
to some setters, mostly the ones that have a chance of getting called rapidly
enough for this to manifest.