gpu: Implement Intel GPU support
Based on an initial implementation from Honjow, add support for Intel GPUs
using both the i915 kernel driver and the Xe kernel driver.

Both i915 and Xe are included in a single "Intel" driver type in order
to support DG2 and older Intel GPUs, which default to i915 but are capable
of using Xe on an experimental basis. If a user switches from i915 to Xe
on a supported device, the config file will not need to be changed.