Add WifiBackend1 DBus interface for WifiBackend property.
Allows WifiBackend to be get and set on non gallileo platforms.
Since most of WifiDebug1 interface is specific to Gallileo devices
this needed to be duplicated to another DBus interface.
Since it doesn't keep any state (besides what's in the config
file on disk) it's pretty straightforward, using the same wifi.rs
methods WifiDebug1 does.